examples/django_demo/requirements.txt
django>=4.0
djangorestframework>=3.14
